--- app/globals.css
@@ -1,34 +1,0 @@
-* {
-    box-sizing: border-box;
-    padding: 0;
-    margin: 0;
-}
-
-title,
-h1,
-h2,
-p {
-    color: #f1f1f1;
-}
-
-html,
-body {
-    max-width: 100vw;
-    overflow-x: hidden;
-    color: #000;
-    background: #000;
-    display: flex;
-    align-items: center;
-}
-
-#companyListItem {
-    text-decoration-color: #f1f1f1;
-    max-width: 100vw;
-    overflow-x: hidden;
-    background-color: #4d4b4b;
-    margin: 5px;
-    padding: 4px;
-    align-items: baseline;
-    display: flex;
-    align-items: center;
-}
