+++ pages/layout.tsx
@@ -1,4 +1,4 @@
-import "./globals.css";
+
 
 export default function RootLayout({
   children,
